Javascript: fix a race condition during parallel setup
